v1.2.0: 影杀 AOE 斩杀 + 嗜血直写回血 + 10 个新配置项 + 代码质量优化
- 影杀觉醒 AOE: 归零时触发斩杀链（之前只削不杀），保留 totalHP 快照防 maxHealth 变化污染
- 嗜血: heal() → HealthUtil.setAllHealthLikeRaw 直写，绕过外部模组 heal() HEAD cancel 拦截（如泽林变体）
- 破敌之眼/淬魂/影杀/嗜血/鸿运当头/受击坚韧/休养生息/不动如山/拒绝篡改: 17 处硬编码 → 10 个 TOML 配置项
- 代码质量: 提取 cleanupExpiredShadowData/executeShadowKill 消除重复，统一类型转换风格

+++ src/main/java/com/ayin90723/adventure_power/ability/ShadowKillAbility.java
@@ -26,7 +26,8 @@ public class ShadowKillAbility implements Ability {
     }
 
     /**
-     * value 返回固伤值（另一值 hpRatio 由 handler 读取）。
+     * 影杀是双参数能力（固伤 + 比例），handler 直接调用 {@link #flatDamage()} 和
+     * {@link #hpRatio()} 获取两个配置值。{@code value()} 仅满足接口契约，实际不被调用。
      */
 
     private int countAtUnlock = 10;
@@ -48,6 +49,6 @@ public class ShadowKillAbility implements Ability {
 
     /** 每次攻击额外削减目标最大生命值的比例，从配置读取 */
     public float hpRatio() {
-        return (float)(double)ModConfig.SHADOW_KILL_HP_RATIO.get();
+        return ModConfig.SHADOW_KILL_HP_RATIO.get().floatValue();
     }
 }
